More 16/32 bit cleanup.

+++ sbs/vmemodules/CADC2530.h
@@ -29,6 +29,7 @@
 #include <signal.h>
 #include <sys/ioctl.h>
 #include <string>
+#include <stdint.h>
 
 #ifndef __CVMEINTERFACE_H
 #include <CVMEInterface.h>
@@ -52,13 +53,13 @@ class CADC2530 {
   // ------------------------------------------------------------
   // Public members
   public:
-    CADC2530(int crateNum = 0, long nBase = 0); //!< 'normal' constructor.
+    CADC2530(int crateNum = 0, uint32_t nBase = 0); //!< 'normal' constructor.
     CADC2530(const CADC2530& card); //!< Copy Constructor.
     ~CADC2530();		//!< Destructor.
 
-    static bool checkCard(int,long,unsigned short&,unsigned short&);
-    static unsigned short volt2lld(double);
-    static unsigned short volt2uld(double);
+    static bool checkCard(int,uint32_t,uint16_t&,uint16_t&);
+    static uint16_t volt2lld(double);
+    static uint16_t volt2uld(double);
 
     CADC2530& operator=(const CADC2530&); //!< Assignment operator
 
@@ -67,8 +68,8 @@ class CADC2530 {
     int readListEvents(void*,int&);      //!< Read multiple events into buffer.
     int readHistogramChannel(void*,int); //!< Read a histogram channel.
 
-    unsigned short cardType();		//!< Return module device type.
-    unsigned short manufacturerId();    //!< Return module Id of card.
+    uint16_t cardType();		//!< Return module device type.
+    uint16_t manufacturerId();    //!< Return module Id of card.
 
     void clearHistogramMemory();        //!< Zero only histogram memory.
     void clearMemory();                 //!< Zero all histo/list memory.
@@ -106,28 +107,28 @@ class CADC2530 {
     void enableSlidingScale();          //!< Enable sliding scale.
     void disableSlidingScale();         //!< Disable sliding scale.
 
-    void setIPL(unsigned short);        //!< Set the interrupt priority level
-    unsigned short getIPL();            //!< Get the interrupt priority level
+    void setIPL(uint16_t);        //!< Set the interrupt priority level
+    uint16_t getIPL();            //!< Get the interrupt priority level
 
-    unsigned short getCSR();            //!< Get the CSR value.
+    uint16_t getCSR();            //!< Get the CSR value.
 
     void dataReadyOnEvent();            //!< Data ready when at least 1 event.
     void dataReadyOnFullness();         //!< Data ready on fullness flags.
 
-    void setLLD(unsigned short);        //!< Set the LLD value.
-    unsigned short getLLD();            //!< Get the LLD value.
-    void setULD(unsigned short);        //!< Set the ULD value.
-    unsigned short getULD();            //!< Get the ULD value.
+    void setLLD(uint16_t);        //!< Set the LLD value.
+    uint16_t getLLD();            //!< Get the LLD value.
+    void setULD(uint16_t);        //!< Set the ULD value.
+    uint16_t getULD();            //!< Get the ULD value.
 
-    void setInterruptVector(unsigned short); //!< Set the interrupt vector. 
-    unsigned short getInterruptVector();     //!< Get the interrupt vector.
-    void setInterruptMask(unsigned short);   //!< Set the interrupt mask.
-    unsigned short getInterruptMask();       //!< Get the interrupt mask.
+    void setInterruptVector(uint16_t); //!< Set the interrupt vector. 
+    uint16_t getInterruptVector();     //!< Get the interrupt vector.
+    void setInterruptMask(uint16_t);   //!< Set the interrupt mask.
+    uint16_t getInterruptMask();       //!< Get the interrupt mask.
 
     void clearFullnessFlags();          //!< Clear the fullness flags to 0.
-    unsigned short getFullnessFlags();  //!< Get the fullness flags.
-    bool isChannelFull(unsigned short); //!< Check if histo. chan. is full.
-    bool channelHasData(unsigned short); //!< Check if histo. channel has data.
+    uint16_t getFullnessFlags();  //!< Get the fullness flags.
+    bool isChannelFull(uint16_t); //!< Check if histo. chan. is full.
+    bool channelHasData(uint16_t); //!< Check if histo. channel has data.
     bool isListFull();                  //!< Check if list full.
     bool isListHalfFull();              //!< Check if list is half full.
 
@@ -135,34 +136,34 @@ class CADC2530 {
     void clearEventCounter();           //!< Clear the event counter. 
 
     // Calibration and test methods
-    unsigned short getCTR();            //!< Get the CTR value.
+    uint16_t getCTR();            //!< Get the CTR value.
     void resetCTR();                    //!< Reset the CTR to zero.
-    void setCTRchannel(unsigned short); //!< Set the CTR channel.
-    unsigned short getCTRchannel();     //!< Get the CTR channel.
+    void setCTRchannel(uint16_t); //!< Set the CTR channel.
+    uint16_t getCTRchannel();     //!< Get the CTR channel.
     void enableAutoFastClear();         //!< Enable CTR auto fast clear. 
     void disableAutoFastClear();        //!< Disable CTR auto fast clear. 
     void enableMUXswitch();             //!< Enable CTR MUX switch.
     void disableMUXswitch();            //!< Disable CTR MUX switch.
     void enableCompensation();          //!< Enable CTR compensation.
     void disableCompensation();         //!< Disable CTR compensation.
-    void setSSTR(unsigned short);       //!< Set the sliding scale test reg. 
-    unsigned short getSSTR();           //!< Get the sliding scale test reg. 
+    void setSSTR(uint16_t);       //!< Set the sliding scale test reg. 
+    uint16_t getSSTR();           //!< Get the sliding scale test reg. 
 
   // ------------------------------------------------------------
   // Protected members
   protected:
     int    my_nCrate;		  //!< VME crate number housing the module.
-    unsigned long my_nBase;	  //!< Base physical VME address in crate.
-    unsigned long my_nMemOffset;  //!< Register vale for list/histogram offset.
-    unsigned short my_nCardId;    //!< Card Id (filled in at MapCard)
-    unsigned short my_nCardType;  //!< Type of card (filled in at MapCard)
+    uint32_t my_nBase;	  //!< Base physical VME address in crate.
+    uint32_t my_nMemOffset;  //!< Register vale for list/histogram offset.
+    uint16_t my_nCardId;    //!< Card Id (filled in at MapCard)
+    uint16_t my_nCardType;  //!< Type of card (filled in at MapCard)
     void*  my_nModFd;		  //!< File desc. for registers open on VME.
     void*  my_nMemFd;		  //!< File desc. for memory open on VME.
     bool my_eventmode;            //!< True when in event mode.
-    unsigned int my_cureventpos;  //!< The current event long position.
+    unsigned int my_cureventpos;  //!< The current event uint32_t position.
 
 
-    volatile unsigned short *my_pModule; //!< Pointer to modulre registers etc.
+    volatile uint16_t *my_pModule; //!< Pointer to modulre registers etc.
     volatile unsigned int *my_pMemory; //!< Pointer to list/histogram memory.
 
 
@@ -178,16 +179,16 @@ class CADC2530 {
 
     int readListEvent(void*,int); //!< Read event into a user buffer.
 
-    unsigned long calcMemoryOffset(unsigned long); //!< Compute memory offset
-    unsigned long setMemoryOffset(unsigned long); //!< Set the memory offset
+    uint32_t calcMemoryOffset(uint32_t); //!< Compute memory offset
+    uint32_t setMemoryOffset(uint32_t); //!< Set the memory offset
 
-    void setCSR(unsigned short);              //!< Set the CSR to a value.
-    void setCSRbits(unsigned short,unsigned short); //!< Set bits in the CSR.
-    void resetCSRbits(unsigned short);        //!< Reset bits in the CSR.
+    void setCSR(uint16_t);              //!< Set the CSR to a value.
+    void setCSRbits(uint16_t,uint16_t); //!< Set bits in the CSR.
+    void resetCSRbits(uint16_t);        //!< Reset bits in the CSR.
 
-    void setCTR(unsigned short);              //!< Set the CTR to a value.
-    void setCTRbits(unsigned short,unsigned short); //!< Set bits in the CTR.
-    void resetCTRbits(unsigned short);        //!< Reset bits in the CTR.
+    void setCTR(uint16_t);              //!< Set the CTR to a value.
+    void setCTRbits(uint16_t,uint16_t); //!< Set bits in the CTR.
+    void resetCTRbits(uint16_t);        //!< Reset bits in the CTR.
 };
 
 #endif
